ARCH-1: extract links dispatch arm into commands::links::run (iter-225)
