+++ crates/hyalo-cli/src/dispatch.rs
@@ -3,7 +3,7 @@ use std::path::Path;
 use anyhow::{Context, Result};
 
 use crate::cli::args::{
-    ChangelogAction, Commands, IndexFlags, LinksAction, LintRulesAction, MadrAction,
+    ChangelogAction, Commands, IndexFlags, LintRulesAction, MadrAction,
     OkfAction, PropertiesAction, TagsAction, TypesAction,
 };
 use crate::commands::inputs::{ResolutionPolicy, ResolvedInputsOrOutcome, resolve_inputs};
@@ -977,140 +977,10 @@ pub(crate) fn dispatch(command: Commands, ctx: &mut CommandContext<'_>) -> Resul
             effective_format,
             allow_outside_vault,
         ),
-        Commands::Links { action } => match action.unwrap_or(LinksAction::Fix {
-            dry_run: true,
-            apply: false,
-            threshold: 0.8,
-            apply_fuzzy: false,
-            min_confidence: None,
-            glob: vec![],
-            ignore_target: vec![],
-            expand_short_form: false,
-            index_flags: IndexFlags::default(),
-        }) {
-            LinksAction::Fix {
-                dry_run: _,
-                apply,
-                threshold,
-                apply_fuzzy,
-                min_confidence,
-                glob,
-                ignore_target,
-                expand_short_form,
-                index_flags: _, // consumed in run.rs before dispatch
-            } => {
-                // Scope the immutable borrow of snapshot_index (via resolve_index)
-                // so we can borrow it mutably for index updates afterwards.
-                let (outcome, modified_files, had_failures) = match resolve_index(
-                    snapshot_index.as_ref(),
-                    dir,
-                    &[],
-                    &[],
-                    effective_format,
-                    site_prefix,
-                    true,
-                    &ScanOptions {
-                        scan_body: true,
-                        bm25_tokenize: false,
-                        default_language: None,
-                        frontmatter_link_props: ctx.frontmatter_link_props,
-                    },
-                )? {
-                    IndexResolution::Resolved(resolved) => {
-                        // `links fix` is entirely about link resolution.
-                        let ci = maybe_case_index(
-                            ctx.case_insensitive_mode,
-                            dir,
-                            true,
-                            resolved.as_snapshot(),
-                        );
-                        links_commands::links_fix(
-                            resolved.as_index(),
-                            dir,
-                            site_prefix,
-                            &glob,
-                            !apply,
-                            threshold,
-                            &ignore_target,
-                            effective_format,
-                            ci.as_ref(),
-                            expand_short_form,
-                            links_commands::FuzzyApply {
-                                apply_fuzzy,
-                                min_confidence,
-                                config_min_confidence: ctx.config_fuzzy_min_confidence,
-                            },
-                        )?
-                    }
-                    IndexResolution::Outcome(outcome) => (outcome, Vec::new(), false),
-                };
-                // L-11: a mid-batch write failure yields a non-zero exit code
-                // even though the envelope is emitted in full.
-                if had_failures {
-                    ctx.exit_code_override = Some(1);
-                }
-                // resolved is dropped — safe to borrow snapshot_index mutably.
-                patch_index_for_modified_files(snapshot_index, index_path, dir, &modified_files)?;
-                Ok(outcome)
-            }
-            LinksAction::Auto {
-                dry_run: _,
-                apply,
-                min_length,
-                exclude_title,
-                first_only,
-                no_first_only,
-                exclude_target_glob,
-                no_warn_common_titles,
-                file,
-                glob,
-                index_flags: _, // consumed in run.rs before dispatch
-            } => {
-                let (outcome, modified_files, had_failures) = match resolve_index(
-                    snapshot_index.as_ref(),
-                    dir,
-                    &[],
-                    &[],
-                    effective_format,
-                    site_prefix,
-                    true,
-                    &ScanOptions {
-                        scan_body: false,
-                        bm25_tokenize: false,
-                        default_language: None,
-                        frontmatter_link_props: ctx.frontmatter_link_props,
-                    },
-                )? {
-                    IndexResolution::Resolved(resolved) => links_commands::links_auto(
-                        resolved.as_index(),
-                        dir,
-                        apply,
-                        &links_commands::AutoFilters {
-                            min_length,
-                            cli_exclude_titles: &exclude_title,
-                            cli_exclude_target_globs: &exclude_target_glob,
-                            cli_first_only: first_only,
-                            cli_no_first_only: no_first_only,
-                            config_exclude_titles: ctx.auto_link_exclude_titles,
-                            config_exclude_target_globs: ctx.auto_link_exclude_target_globs,
-                            config_first_only: ctx.auto_link_first_only,
-                            cli_no_warn_common_titles: no_warn_common_titles,
-                            config_warn_common_titles: ctx.auto_link_warn_common_titles,
-                        },
-                        file.as_deref(),
-                        &glob,
-                        effective_format,
-                    )?,
-                    IndexResolution::Outcome(outcome) => (outcome, Vec::new(), false),
-                };
-                // L-11: partial write failure ⇒ non-zero exit code.
-                if had_failures {
-                    ctx.exit_code_override = Some(1);
-                }
-                patch_index_for_modified_files(snapshot_index, index_path, dir, &modified_files)?;
-                Ok(outcome)
-            }
-        },
+        Commands::Links { action } => {
+            // ARCH-1 (iter-225): the arm body now lives in `commands::links::run`.
+            links_commands::run(ctx, action)
+        }
         Commands::Lint {
             file_positional,
             file,
